fix: Fixed an issue where the program failed to run properly when compiled with -O0
Fix the issue where using `static` for UART results in different copies
of the file

Added the setting of `mtimecmp` during clock interrupt initialization to
prevent the system from suspending immediately after the clock interrupt
is enabled.

+++ arch/riscv/boot/mstart.c
@@ -29,7 +29,7 @@ __attribute__((noreturn)) void mstart(void)
 	// Force refresh
 	sfence_vma();
 
-	int hartid = r_mhartid();
+	int hartid = (int) r_mhartid();
 	w_mscratch((uint64) &mscratch0[hartid * 32]);
 	w_mtvec((uint64) m_trap_handler);
 
@@ -57,7 +57,7 @@ __attribute__((noreturn)) void mstart(void)
 	//           I_S_STORE_PAGE_FAULT);
 
 	// keep each CPU's hartid in its tp register, for cpuid().
-	int id = r_mhartid();
+	int id = (int) r_mhartid();
 	w_tp(id);
 
 	// set the starting position of the MEPC
